Store jwt in res cookie
I'm transitioning the auth approach to storing jwt access token in res
cookie instead of sending it to client to be stored in localStorage.
This also refactors the middlewares and utils used for authorization.
Instead of saving session tokens to db and checking it for each request,
I'm creating a short-lived access token with associated
refresh token that's saved on the db. DB lookups only happens when the
access token has expired. Refresh tokens have longer expiration time of
15 days, which I think is reasonable.